+++ src/main/java/ru/vyarus/dropwizard/guice/test/GuiceyHooksRule.java
@@ -3,7 +3,7 @@ package ru.vyarus.dropwizard.guice.test;
 import org.junit.rules.ExternalResource;
 import ru.vyarus.dropwizard.guice.hook.GuiceyConfigurationHook;
 import ru.vyarus.dropwizard.guice.hook.ConfigurationHooksSupport;
-import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyConfiguration;
+import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyHooks;
 
 import java.util.Arrays;
 import java.util.List;
@@ -16,13 +16,13 @@ import java.util.List;
  * <pre>{@code static GuiceyAppRule RULE = new GuiceyAppRule(App.class, null);
  *    {@literal @}ClassRule
  *    public static RuleChain chain = RuleChain
- *            .outerRule(new GuiceyConfigurationRule((builder) -> builder.modules(...)))
+ *            .outerRule(new GuiceyHooksRule((builder) -> builder.modules(...)))
  *            .around(RULE);
  * }</pre>
  * To declare common extensions for all tests, declare common rule in test class (without {@code @ClassRule}
  * annotation!) and use it in chain:
  * <pre>{@code public class BaseTest {
- *         static GuiceyConfigurationRule BASE = new GuiceyConfigurationRule((builder) -> builder.modules(...))
+ *         static GuiceyHooksRule BASE = new GuiceyHooksRule((builder) -> builder.modules(...))
  *     }
  *
  *     public class SomeTest extends BaseTest {
@@ -30,13 +30,13 @@ import java.util.List;
  *         {@literal @}ClassRule
  *         public static RuleChain chain = RuleChain
  *            .outerRule(BASE)
- *            .around(new GuiceyConfigurationRule((builder) -> builder.modules(...)) // optional test-specific staff
+ *            .around(new GuiceyHooksRule((builder) -> builder.modules(...)) // optional test-specific staff
  *            .around(RULE);
  *     }
  * }</pre>
  * <p>
  * IMPORTANT: rule will not work with spock extensions (because of lifecycle specifics)! Use
- * {@link UseGuiceyConfiguration} or new {@code hooks} attribute in
+ * {@link UseGuiceyHooks} or new {@code hooks} attribute in
  * {@link ru.vyarus.dropwizard.guice.test.spock.UseGuiceyApp} or
  * {@link ru.vyarus.dropwizard.guice.test.spock.UseDropwizardApp} instead.
  * <p>
@@ -45,11 +45,11 @@ import java.util.List;
  * @author Vyacheslav Rusakov
  * @since 11.04.2018
  */
-public class GuiceyConfigurationRule extends ExternalResource {
+public class GuiceyHooksRule extends ExternalResource {
 
     private final List<GuiceyConfigurationHook> hooks;
 
-    public GuiceyConfigurationRule(final GuiceyConfigurationHook... hooks) {
+    public GuiceyHooksRule(final GuiceyConfigurationHook... hooks) {
         this.hooks = Arrays.asList(hooks);
     }